moved Routine Activities & Fragments to /routine; moved Video Activity & Fragments to /video
