+++ MyFitnessRoutines/src/main/java/com/devindow/myfitnessroutines/SectionsPagerAdapter.java
@@ -6,6 +6,9 @@ import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentPagerAdapter;
 
+import com.devindow.myfitnessroutines.routine.PracticeFragment;
+import com.devindow.myfitnessroutines.video.LearnFragment;
+
 /**
  * A [FragmentPagerAdapter] that returns a fragment corresponding to
  * one of the sections/tabs/pages.
