+++ MyFitnessRoutines/src/main/java/com/devindow/myfitnessroutines/routine/MainActivity.java
@@ -1,4 +1,4 @@
-package com.devindow.myfitnessroutines;
+package com.devindow.myfitnessroutines.routine;
 
 import android.app.AlertDialog;
 import android.content.Context;
@@ -16,9 +16,13 @@ import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import com.devindow.myfitnessroutines.BuildConfig;
+import com.devindow.myfitnessroutines.OptionsMenuActivity;
+import com.devindow.myfitnessroutines.R;
 import com.devindow.myfitnessroutines.db.AppDatabase;
 import com.devindow.myfitnessroutines.routine.*;
 import com.devindow.myfitnessroutines.util.MethodLogger;
+import com.devindow.myfitnessroutines.video.VideoStreamActivity;
 
 import java.util.List;
