+++ MyFitnessRoutines/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.devindow.myfitnessroutines.MainActivity">
+    tools:context="com.devindow.myfitnessroutines.routine.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
