+++ MyFitnessRoutines/src/main/res/layout/content_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.devindow.myfitnessroutines.MainActivity"
+    tools:context="com.devindow.myfitnessroutines.routine.MainActivity"
     tools:showIn="@layout/activity_main">
 
     <ListView
